Default Version

+++ src/test/java/edu/spbu/MatrixGeneratorTest.java
@@ -11,31 +11,30 @@ import java.io.FileReader;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-public class MatrixGeneratorTest
-{
-  String fileName="test.txt";
-  @After
-  public void cleanUp() {
-    File f = new File(fileName);
-    if(f.exists())
-      f.delete();
-  }
+public class MatrixGeneratorTest{
+    String fileName="test.txt";
 
-  @Test
-  public void testGenerate() throws Exception {
-    new MatrixGenerator(1,3,fileName,10).generate();
-    BufferedReader reader = new BufferedReader(new FileReader(fileName));
-    int lineCount=0;
-    int emptyLineCount=0;
-    for (String line = reader.readLine();line !=null; line = reader.readLine()) {
-      if (line.equals("0 0 0 0 0 0 0 0 0 0")) {
-        emptyLineCount++;
-      } else {
-        assertEquals(10, line.split(" ").length);
-      }
-      lineCount++;
+    @After
+    public void cleanUp(){
+        File f=new File(fileName);
+        if(f.exists()) f.delete();
     }
-    assertTrue(emptyLineCount>1);
-    assertEquals(10,lineCount);
-  }
+
+    @Test
+    public void testGenerate() throws Exception{
+        new MatrixGenerator(1,3,fileName,10).generate();
+        BufferedReader reader=new BufferedReader(new FileReader(fileName));
+        int lineCount=0;
+        int emptyLineCount=0;
+        for(String line=reader.readLine();line!=null;line=reader.readLine()){
+            if(line.equals("0 0 0 0 0 0 0 0 0 0")){
+                emptyLineCount++;
+            }else{
+                assertEquals(10,line.split(" ").length);
+            }
+            lineCount++;
+        }
+        assertTrue(emptyLineCount>1);
+        assertEquals(10,lineCount);
+    }
 }
